chore: move instance command generation service
