+++ packages/twenty-server/src/engine/core-modules/upgrade/upgrade.module.ts
@@ -3,7 +3,6 @@ import { DiscoveryModule } from '@nestjs/core';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { InstanceUpgradeEntity } from 'src/engine/core-modules/upgrade/instance-upgrade.entity';
-import { InstanceCommandGenerationService } from 'src/engine/core-modules/upgrade/services/instance-command-generation.service';
 import { InstanceUpgradeService } from 'src/engine/core-modules/upgrade/services/instance-upgrade.service';
 import { RegisteredInstanceMigrationService } from 'src/engine/core-modules/upgrade/services/registered-instance-migration-registry.service';
 import { WorkspaceUpgradeService } from 'src/engine/core-modules/upgrade/services/workspace-upgrade.service';
@@ -17,15 +16,11 @@ import { WorkspaceEntity } from 'src/engine/core-modules/workspace/workspace.ent
   providers: [
     InstanceUpgradeService,
     WorkspaceUpgradeService,
-    // TODO prastoin does not make really sense to be exported here
-    InstanceCommandGenerationService,
     RegisteredInstanceMigrationService,
   ],
   exports: [
     InstanceUpgradeService,
     WorkspaceUpgradeService,
-    // TODO prastoin does not make really sense to be exported here
-    InstanceCommandGenerationService,
     RegisteredInstanceMigrationService,
   ],
 })
